Ruby is dead, long live Python
This commit kills our previous Ruby/Sinatra application and adds a
bare-bones Python/Flask application instead. The justification for this
is that the group as a whole is more familiar with Python, so this will
make the app more maintainable going forward.

Right now, there's no DB, external data, or anything. The
`api/hackathon` endpoint exists to provide a temporary API for the
front-end app to work against.

Next up:
  - Attach a database
  - Add some API safeguards/conveniences like versioning and caching
  - Generate API docs?