Remove editable mode dependency
Fix the build.

I installed progress_report in editable mode (`pip install -e .`) in
order to allow `pytest` to find the project root. This breaks both
Travis and the Heroku build, and it's not a good idea for production.

This commit removes that dependency and changes Travis's test script to
`python3 -m pytest`; that should allow pytest to load the module without
including an editable dependency.