Use sqlite for integration/datastore tests
The goal of this commit is to get Travis to pass.

This commit introduces a fair amount of technical debt. The reasoning is
that until we add database migrations (and possibly figure out our
long-term table structure), it'll be difficult to test with postgresql.
Because of this, this commit introduces an in-memory sqlite database for
testing.